Invert model between history and memory strategy
